--- app/src/androidTest/java/org/mozilla/fenix/ui/SettingsLoginsAndPasswordsTest.kt
@@ -1,51 +1,0 @@
-/* This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
-
-package org.mozilla.fenix.ui
-
-import androidx.test.platform.app.InstrumentationRegistry
-import androidx.test.uiautomator.UiDevice
-import org.junit.Rule
-import org.junit.Test
-import org.mozilla.fenix.helpers.HomeActivityTestRule
-import org.mozilla.fenix.helpers.TestHelper
-import org.mozilla.fenix.ui.robots.homeScreen
-
-/**
- *  Tests for verifying the Logins and Passwords settings menu
- *
- */
-
-class SettingsLoginsAndPasswordsTest {
-
-    private val mDevice = UiDevice.getInstance(InstrumentationRegistry.getInstrumentation())
-
-    @get:Rule
-    val activityTestRule = HomeActivityTestRule()
-
-    // Tests only for initial state without signing in.
-    // For tests after singing in, see SyncIntegration test suite
-
-    @Test
-    fun loginsAndPasswordsTest() {
-        homeScreen {
-        }.openThreeDotMenu {
-        }.openSettings {
-            // Necessary to scroll a little bit for all screen sizes
-            TestHelper.scrollToElementByText("Logins and passwords")
-        }.openLoginsAndPasswordSubMenu {
-            verifyDefaultView()
-            verifyDefaultValueSyncLogins()
-        }.openSavedLogins {
-            verifySavedLoginsView()
-            tapSetupLater()
-            // Verify that logins list is empty
-            // Issue #7272 nothing is shown
-        }.goBack {
-        }.openSyncLogins {
-            verifyReadyToScanOption()
-            verifyUseEmailOption()
-        }
-    }
-}
